fleshes out the example more
