+++ tests/test_app.py
@@ -3,7 +3,7 @@ This module contains tests for the FastAPI application.
 """
 
 from fastapi.testclient import TestClient
-from ..app import app
+from ..app_api import app
 
 client = TestClient(app)
